Generate complete H2 controller BSP maps

+++ targets/hub_rp/main.c
@@ -14,7 +14,7 @@ int main(void)
     for (;;) {
         if (image_identity != L2_HUB_RP_IMAGE_IDENTITY ||
             l2_r2_hub_rp_domain.mapping != L2_R2_MAPPING_EXACT_PINS ||
-            l2_r2_hub_rp_domain.pin_count != UINT8_C(48)) {
+            l2_r2_hub_rp_domain.pin_count != UINT16_C(48)) {
             return 1;
         }
         tight_loop_contents();
